requirements.txt
selenium>=4.10.0
requests>=2.28.0
python-dotenv>=1.0.0

# For FREE audio CAPTCHA bypass (optional but recommended)
SpeechRecognition>=3.10.0
pydub>=0.25.1

# For audio processing (required for pydub)
# Install ffmpeg separately: apt-get install ffmpeg (Linux) or brew install ffmpeg (Mac)
